numero decimal

+++ paginas/models.py
@@ -6,7 +6,7 @@ class tb_balanco(models.Model):
     codigo = models.CharField(max_length=50)
     descricao = models.CharField(max_length=150, verbose_name='Descrição')
     familia = models.CharField(max_length=50, verbose_name='Lista')
-    qt = models.IntegerField()
+    qt = models.FloatField()
 
     class Meta:
         db_table = 'paginas_tb_balanco'
